Rename pipPackageManager.ts
